backend/requirements.txt
Flask==3.0.3
firebase-admin==6.6.0
PyJWT==2.9.0
gunicorn==23.0.0
